+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - walk inheritance chain to find implementation
 - fix: nil should not inherit Object implementations
 - fix: walk superclass chain even with :extend-via-metadata
+- a value always implements an :extend-via-methdata protocol with zero methods
 
 # 1.0.0 - 10th September 2021
 - initial public release